Use CFG's internal region on CFGNodes
Base class CFGNode has two constructors,
1) one that takes a TR_Memory * parameter
2) one that takes a TR::Region& parameter
When the first constructor is used the sucessors and predecessors
are stored in the parameter region. Otherwise, they are stored in the
heap memory region.
Child class OMR::Block only implements the first type of constructor.
This commit creates another constructor for OMR::Block which takes a
TR::CFG & parameter and uses the internal memory region for calling
the second type of CFGNode constructor.

Signed-off-by: Erick Ochoa <eochoa@ualberta.ca>